+++ Breve/Classes/View Controllers/BKBreveSettingsViewController.m
@@ -90,7 +90,7 @@
         }
     }
     BKBreveSettings *settings = [BKBreveSettings generalSettings];
-    [settings setTheme:indexPath.row];
+    [settings setTheme:(BKBreveTheme)indexPath.row];
 
     [self applyTheme];
 }